fix(1085): extract the responsive helper so specs don't import specs
Playwright refuses to load a suite where one spec file imports another, so
pulling expectNoHorizontalPageScroll from responsive.spec.ts failed every E2E
shard at collection time — not a flake, a real break I introduced.

The helper belongs in helpers/ regardless: the mobile guard is meant to be
reusable, so any spec adding a surface can assert the invariant without
duplicating the check.

Part of #960
